解决页面间跳转存在问题，实现表tSFB数据分页，条件查询，默认获取30天内数据 创建触发器，在录入员工信息同时一并录入登录信息，初始密码8个8，code自增未实现
实现code自增，删除员工信息同时一并删除登录信息，用户冻结以及启用

实现未登录跳转到登录界面，前端界面间传送数据，未实现登录用户身份判断，有些功能只能管理员操作

实现登录用户身份判断,管理员才可以操作相关功能，收费录入存在问题

完成收费录入，收费项目完成，尚未完成【车辆监管录入】，【检测发证】，【登出功能】

【车辆监管录入】未完成，【检测发证】界面完成，数据表格待处理，【检测项目单项耗时计算完成】，【检测项目单项完成结束，开始触发未完成】

+++ src/main/java/com/boot/damain/chargeInfo.java
@@ -22,7 +22,7 @@ public class chargeInfo {
     private String memo;    //备注信息
     private String jcxm;     //检测项目
     private int zsl;        //总数量
-    private int zje;        //总金额
+    private Double zje;        //总金额
     private String optname; //操作人
     private String flag;       //
     private String dsdate;    //
